backend/app/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
SQLAlchemy
alembic
pydantic
python-dotenv
email-validator